Migrate breadcrumbs rendering to a ViewComponent
Extract the breadcrumb markup out of the fixed-elements builder into a
per-brand BreadcrumbsComponent (generic base, GOV.UK ordered list, NHS
labelled nav with a back link). The fixed-elements builder still owns
the breadcrumb accumulation API and content_for(:breadcrumbs) slot, but
now renders the brand component instead of building the HTML inline.

- Generic Elements::Breadcrumbs renders Registry.component(brand,
  :breadcrumbs); the GOV.UK/NHS breadcrumb element mixins and their
  FixedElements includes are removed (behaviour moves into the
  component subclasses)
- current_page? aria-current handling preserved via helpers.current_page?
- Verified live for both brands and via the existing breadcrumbs specs
