+++ lib/design_system/generic/builders/elements/breadcrumbs.rb
@@ -2,7 +2,8 @@ module DesignSystem
   module Generic
     module Builders
       module Elements
-        # This mixin module is used to provide breadcrumbs.
+        # This mixin accumulates breadcrumbs and renders them into the
+        # :breadcrumbs content_for slot via the per-brand BreadcrumbsComponent.
         module Breadcrumbs
           def breadcrumb(label, path)
             @breadcrumbs ||= []
@@ -13,30 +14,9 @@ module DesignSystem
 
           def content_for_breadcrumbs
             content_for(:breadcrumbs) do
-              safe_buffer = ActiveSupport::SafeBuffer.new
-
-              @breadcrumbs.each do |breadcrumb|
-                safe_buffer.concat(render_breadcrumb(breadcrumb))
-                safe_buffer.concat(' > ') unless @breadcrumbs.last == breadcrumb
-              end
-
-              safe_buffer
+              component = ::DesignSystem::Registry.component(brand, :breadcrumbs).new(breadcrumbs: @breadcrumbs)
+              @context.render(component)
             end
-          end
-
-          def home_path?(path)
-            root_hash = Rails.application.routes.recognize_path(@context.root_path)
-            path_hash = Rails.application.routes.recognize_path(path)
-
-            %i[controller action].each do |key|
-              return false if root_hash[key] != path_hash[key]
-            end
-
-            true
-          end
-
-          def render_breadcrumb(breadcrumb)
-            link_to_unless_current(breadcrumb[:label], breadcrumb[:path])
           end
         end
       end
