+++ lib/design_system/govuk/builders/fixed_elements.rb
@@ -5,7 +5,6 @@ module DesignSystem
     module Builders
       # This class is used to provide the GOVUK fixed elements builder.
       class FixedElements < ::DesignSystem::Generic::Builders::FixedElements
-        include Elements::Breadcrumbs
         include Elements::Headings
       end
     end
